requirements.txt
fabric==1.6.1
boto==2.27.0
PyYAML==3.11